+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to the Aptos Python SDK will be captured in this file. This 
 
 ## Unreleased
 - Add Multikey support for Python, with an example
-- Deprecate non-BCS transaction submission
+- Deprecate and remove non-BCS transaction submission
 
 ## 0.8.6
 - add client for graphql indexer service with light demo in coin transfer